Rename lib => src
