+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "start": "cross-env NODE_ENV=development webpack-dev-server",
     "build": "cross-env NODE_ENV=production webpack -p",
-    "compile": "babel lib --out-dir dist",
+    "compile": "babel src --out-dir dist",
     "lint": "eslint .",
     "lint-fix": "eslint . --fix",
     "codecov": "codecov --disable=gcov",
