+++ src/Terminal.jsx
