+++ src/defs/types/Terminal.js
