+++ src/defs/types/TerminalMessage.js
