+++ src/handlers/parseEOL.js
