+++ src/handlers/validateCommands.js
