+++ src/utils/cleanArray.js
